backend/requirements.txt
fastapi
uvicorn
pydantic
email-validator
pymongo
python-dotenv
passlib[bcrypt]
python-jose
